# Backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EVADIA API",
        "description": "API complète de la plateforme de gestion hôtelière EVADIA. Comprend l'authentification, la gestion admin, le back-office hôtel et les endpoints client.",
        "contact": {
            "name": "EVADIA",
            "email": "contact@evadia.com"
        },
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/api/admin/hotels/{id}": {
            "get": {
                "tags": [
                    "Admin - Hôtels"
                ],
                "summary": "Détails d'un hôtel",
                "description": "Récupère les informations complètes d'un hôtel spécifique",
                "operationId": "833bd6182e677d442a1e3aa951c075e1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'hôtel",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Informations de l'hôtel",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Hôtel non trouvé"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Admin - Hôtels"
                ],
                "summary": "Mettre à jour un hôtel",
                "description": "Modifie les données d'un hôtel existant.",
                "operationId": "c7bf9929424688ed031a0f9fdc69eb63",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'hôtel",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Hôtel Paris Modifié"
                                    },
                                    "email_contact": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "contact@hotel-paris.com"
                                    },
                                    "telephone": {
                                        "type": "string",
                                        "example": "+33111222333"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Description de l'hôtel modifiée"
                                    },
                                    "etoiles": {
                                        "type": "integer",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Hôtel mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Hôtel mis à jour avec succès"
                                        },
                                        "hotel": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Hôtel non trouvé"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin - Hôtels"
                ],
                "summary": "Supprimer un hôtel",
                "description": "Supprime un hôtel de la base de données.",
                "operationId": "0f225a99b3e6492358b109932339d719",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'hôtel",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hôtel supprimé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Hôtel supprimé avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Hôtel non trouvé"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/client/hotels/{id}": {
            "get": {
                "tags": [
                    "Client - Hôtels"
                ],
                "summary": "Détails d'un hôtel",
                "description": "Retourne les informations complètes d'un hôtel : description, photos, chambres avec prix, services, avis.",
                "operationId": "5f5ffaf948d6b702bb4784a0532dc206",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de l'hôtel",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "hotel": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "nom": {
                                                    "type": "string",
                                                    "example": "Hôtel Renaissance"
                                                },
                                                "description": {
                                                    "type": "string"
                                                },
                                                "etoiles": {
                                                    "type": "integer",
                                                    "example": 4
                                                },
                                                "email_contact": {
                                                    "type": "string"
                                                },
                                                "telephone": {
                                                    "type": "string"
                                                },
                                                "site_web": {
                                                    "type": "string",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "photos": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "url_photo": {
                                                        "type": "string"
                                                    },
                                                    "legende": {
                                                        "type": "string",
                                                        "nullable": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "chambres": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "nom": {
                                                        "type": "string"
                                                    },
                                                    "type_propriete": {
                                                        "type": "string"
                                                    },
                                                    "capacite": {
                                                        "type": "integer"
                                                    },
                                                    "prix_par_nuit": {
                                                        "type": "number",
                                                        "nullable": true
                                                    },
                                                    "photo": {
                                                        "type": "string",
                                                        "nullable": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "services": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "note_moyenne": {
                                            "type": "number",
                                            "format": "float",
                                            "nullable": true
                                        },
                                        "nb_avis": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Hôtel non trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token obtenu via /api/auth/login ou /api/auth/register",
                "bearerFormat": "Sanctum Token",
                "scheme": "bearer"
            }
        }
    },
...
}